pyproject.toml
[tool.poetry]
name = "ViT-mini"
version = "0.1.0"
description = ""
authors = ["ChrisTho23 <christophe.thomassin23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torchvision = "0.17"
matplotlib = "^3.8.4"
numpy = "^1.26.4"
wandb = "^0.17.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
